Extend gbp.conf and write a README.source to document new upstream import
Configure git-buildpackage permanently via a gbp.conf file to allow
running git-buildpackage commands without having to remember all the
options.

Comment the config file extensively to explain each option. Also add a
README for package maintainers and contributors to make it explicit on
what are the correct commands to run for imports and contributions.