Salsa CI: Replace bullseye-backports with trixie-backports
With the release of Debian 13 "Trixie", the old Debian 11 "Bullseye"
backports has been archived and due to built-in Debian archive selection
in sbuild, we can't override it to use archive.debian.org and it is best
to just remove it. Add Trixie instead as it will be useful going forward.